move sessions and pending sessions to communicator
